First pass at refactoring the workflow code
Changed the workflow_enabled flag on teams to be workflow_style.
There are now 3 options, None, Tasks, or Collaboration.  Collaboration
is the new style that still needs to be implemented.

Moved the workflow-specific code out of teams.models and into
teams.workflow.  There's already enough there that it deserves its own
module and we're about to add more.

Renamed Workflow to "TaskWorkflow".  We're about to add
CollaborationWorkflow to complement it.

Upgraded south.  We needed version 0.8 to handle migrating the unique
indexes.

Changed the way we handle ALL_LANGUAGES.  Trying to translate them the
way we were resulted in us importing other modules in the middle of
importing the subtitles.models modules, which resulted in errors.

Pulled in a new integration repo commit.