Fixed collaboration team workings
Made it so that team tracks the team working on the collaboration and
added some code to make that not always match the owner of team_video.
Allow team to be NULL before the first collaborator gets added.

Made Collaboration.add_collaborator() input a TeamMember rather than
just a user.  This way we can check/assign the Collaboration team in
that method.  Changed mark_endorsed() to input a TeamMember to keep
things consistent.