Moved the collaboration language select form to the workflow page
I think it makes sense here, especially since the workflow page is the
one place we send people to after the team is created.

Also changed the workflow languages to be a radio button + select
combo.  This should make it easier to add more settings like TED's
"create for the primary audio-language, then create for all
languages."