One last rehaul of the dashboard query
Changed it so that collaborations are split up into ones the user has
joined and ones they can join.  For the collaborations the user can
join we:
  - Order them by the collaboration state.
  - Limit the number we return for each state.

I think this is the cleanest way to give a good experience for the
dashboard page.  It should work okay regardless of the
only_1_subtitler/only_1_reviewer/only_1_approver settings.