Started on the collaboration dashboard
List collaborations the user has jonied/can join.  Added collaboration
url/view.  Right now the only thing you can do to it is POST to it to
join it.  It redirects to the subtitle editor with collaboration_id
set, although the editor doesn't yet understand that GET param.

Changed the Collaboration.state values so that they are not the same
but with different case.  MySQL was using a case-insensitive match,
which caused issues.