Fixed round 1 of issues with the collaborations stuff (#1085)
Fixed crash when viewing the dashboard while not logged in

Don't check the tasks enabled setting for teams with the default
workflow.

Removed the Languages tab from the team settings page

Fixed typo with when creating a new collaborations-style team.

Default to english if the user hasn't picked any languages.  Added
languages to the user admin form.

For subtitlers, display the progression sidebar instead of the
collaboration sidebar.

Made the language list work with collaboration-style teams