Implemented UI for collaboration notes (#1089)
It's just filled with nonsense data for now.

Refactored how we implplement the collab sidebar.  We're not using a
lot of the features like expanding/collapsing the sidebar/sections, so
I removed those parts.  This keeps the HTML and javascript feeling
cleaner to me.  Also added as switch that changes the collab section
from working the tasks to working with the new collaboration model.